Better names for workspace providers based on the locking they provide
